requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
awscli==1.35.14
bcrypt==4.0.1
blinker==1.6.2
botocore==1.35.48
certifi==2023.11.17
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
convertdate==2.4.0
Django==4.2.3
docutils==0.16
ephem==4.1.6
fastapi==0.115.12
Flask==2.3.2
frozenlist==1.4.0
gunicorn==21.2.0
h11==0.16.0
holidays==0.66
httptools==0.6.4
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
LunarCalendar==0.0.9
MarkupSafe==2.1.3
multidict==6.0.4
mysql-connector-python==8.1.0
mysqlclient==2.2.1
numpy==2.2.2
openai==0.27.8
packaging==23.1
pandas==2.2.3
pillow==11.2.1
protobuf==4.21.12
pyasn1==0.6.1
pydantic==2.11.4
pydantic_core==2.33.2
pylinktester==0.1
PyMeeus==0.5.12
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
requests==2.31.0
rsa==4.7.2
s3transfer==0.10.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
starlette==0.46.2
tqdm==4.66.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.1
urllib3==2.0.4
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==2.3.6
yarl==1.9.2
